Added some modifies.
From now, there is a vector of thread and I can have a variable number of robots (so rovers), it depends on how many start positions I give to the code.

From this commit, when a robot reaches its goal, it searches for the next one and continues until all goals are consumed.